requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite
backoff==1.8.0
PyMySQL==0.9.2
peewee==3.7.1
aioredis==1.2.0
ruamel.yaml==0.15.88